+++ src/backend/benchmark/logger/logger.cpp
@@ -15,7 +15,7 @@
 
 #include "backend/benchmark/logger/logger.h"
 #include "backend/benchmark/logger/logger_configuration.h"
-#include "logger_workload.h"
+#include "backend/benchmark/logger/logger_workload.h"
 
 namespace peloton {
 namespace benchmark {
@@ -25,6 +25,19 @@ configuration state;
 
 // Main Entry Point
 void RunBenchmark() {
+
+  // Experiment
+  switch (state.experiment_type) {
+    case EXPERIMENT_TYPE_ACTIVE:
+      RunActiveExperiment();
+      break;
+
+    default:
+      std::cout << "Unsupported experiment type : " << state.experiment_type
+      << "\n";
+      break;
+  }
+
 }
 
 }  // namespace logger
@@ -32,8 +45,9 @@ void RunBenchmark() {
 }  // namespace peloton
 
 int main(int argc, char **argv) {
-  peloton::benchmark::logger::ParseArguments(
-      argc, argv, peloton::benchmark::logger::state);
+
+  peloton::benchmark::logger::ParseArguments(argc, argv,
+                                             peloton::benchmark::logger::state);
 
   peloton::benchmark::logger::RunBenchmark();
