+++ tests/logging/Makefile.am
@@ -4,10 +4,4 @@
 # LOGGING
 ######################################################################
 
-check_PROGRAMS += \
-	       logging_test
-
-logging_test_SOURCES = \
-           logging/logging_tests_util.cpp \
-           logging/logging_test.cpp \
-           harness.cpp
+check_PROGRAMS +=
